# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/books": {
      "post": {
        "description": "Used by Authors to create a new book. Books start in a draft state unless submitted for review. Supports multipart/form-data for uploading cover images, audiobooks, PDFs, etc.",
        "operationId": "BooksController_createBook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Book details and optional file uploads (bookCover, audiobook, ebook, hardcover, paperback, interiorPdf, coverPdf)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully"
          }
        },
        "summary": "Create a new book (draft or submit for review)",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      },
      "get": {
        "operationId": "BooksController_getAllBooks",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "ageGroup",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of books"
          }
        },
        "summary": "Get all books with filtering, search, and pagination",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/approved": {
      "get": {
        "operationId": "BooksController_getApprovedBooks",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of approved books"
          }
        },
        "summary": "Get all approved/published books",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/pending": {
      "get": {
        "operationId": "BooksController_getPendingBooks",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending books"
          }
        },
        "summary": "Get all pending (submitted) books for review",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/mine": {
      "get": {
        "operationId": "BooksController_getMyBooks",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated author book list"
          }
        },
        "summary": "Get all books owned by the current author",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/author/{authorId}": {
      "get": {
        "operationId": "BooksController_getBooksByAuthor",
        "parameters": [
          {
            "name": "authorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of books by author"
          }
        },
        "summary": "Get books by author ID",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/{id}": {
      "get": {
        "operationId": "BooksController_getBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book details"
          },
          "404": {
            "description": "Book not found"
          }
        },
        "summary": "Get a single book by ID",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      },
      "put": {
        "operationId": "BooksController_updateBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Book fields to update and optional new file uploads (bookCover, audiobook, ebook, hardcover, paperback, interiorPdf, coverPdf)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          }
        },
        "summary": "Update a draft or rejected book",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      },
      "delete": {
        "operationId": "BooksController_deleteBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          }
        },
        "summary": "Delete a book",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/{id}/submit": {
      "patch": {
        "operationId": "BooksController_submitForReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book submitted for review"
          }
        },
        "summary": "Submit a draft book for review",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
    "/books/{id}/approve": {
      "patch": {
        "operationId": "BooksController_approveBook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book status updated"
          }
        },
        "summary": "Approve or reject a submitted book",
        "tags": [
          "Books",
          "Author",
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "Production-ready NestJS API with Prisma, PostgreSQL, Authentication, Logging, Monitoring, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "BookFormatDto": {
        "type": "object",
        "properties": {
          "formatType": {
            "type": "string",
            "example": "EBOOK"
          },
          "listPrice": {
            "type": "number",
            "example": 9.99
          },
          "sku": {
            "type": "string",
            "example": "SKU-123"
          },
          "pageCount": {
            "type": "number",
            "example": 320
          },
          "trimSize": {
            "type": "string",
            "example": "6x9"
          },
          "coverUrl": {
            "type": "string",
            "example": "https://..."
          },
          "interiorUrl": {
            "type": "string",
            "example": "https://..."
          }
        },
        "required": [
          "formatType",
          "listPrice"
        ]
      },
      "PrintEditionDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "trimSize": {
            "type": "string",
            "example": "6x9"
          },
          "bindingType": {
            "type": "string",
            "example": "Perfect"
          },
          "interiorColor": {
            "type": "string",
            "example": "Black & White"
          },
          "paperType": {
            "type": "string",
            "example": "60# Uncoated White"
          },
          "bookType": {
            "type": "string",
            "example": "Pocketbook"
          },
          "printQuality": {
            "type": "string",
            "example": "Standard"
          },
          "coverFinish": {
            "type": "string",
            "example": "Gloss"
          },
          "authorProfit": {
            "type": "number",
            "example": 4
          },
          "sellingPrice": {
            "type": "number",
            "example": 9.99
          }
        },
        "required": [
          "enabled",
          "trimSize",
          "bindingType",
          "interiorColor",
          "paperType",
          "bookType",
          "printQuality",
          "coverFinish"
        ]
      },
      "CreateBookRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "The Great Adventure"
          },
          "description": {
            "type": "string",
            "example": "A thrilling adventure story..."
          },
          "isbn": {
            "type": "string",
            "example": "978-3-16-148410-0"
          },
          "category": {
            "type": "string",
            "example": "Fiction"
          },
          "tags": {
            "example": [
              "adventure",
              "fiction",
              "thriller"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string",
            "example": "English"
          },
          "ageGroup": {
            "type": "string",
            "example": "Adult"
          },
          "formats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookFormatDto"
            }
          },
          "authorEarnings": {
            "type": "number",
            "example": 7
          },
          "publicationDetails": {
            "type": "string",
            "example": "Published by Wonder Press, 2024"
          },
          "printEdition": {
            "$ref": "#/components/schemas/PrintEditionDto"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateBookRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "The Great Adventure"
          },
          "description": {
            "type": "string",
            "example": "A thrilling adventure story..."
          },
          "isbn": {
            "type": "string",
            "example": "978-3-16-148410-0"
          },
          "category": {
            "type": "string",
            "example": "Fiction"
          },
          "tags": {
            "example": [
              "adventure",
              "fiction"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "language": {
            "type": "string",
            "example": "English"
          },
          "ageGroup": {
            "type": "string",
            "example": "Adult"
          },
          "formats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookFormatDto"
            }
          },
          "authorEarnings": {
            "type": "number",
            "example": 7
          },
          "publicationDetails": {
            "type": "string",
            "example": "Published by Wonder Press, 2024"
          },
          "printEdition": {
            "$ref": "#/components/schemas/PrintEditionDto"
          }
        }
      },
      "ApproveBookRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "APPROVED",
              "REJECTED"
            ],
            "example": "APPROVED"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}